Refactor TapSigner backup export logic
